Make sidecar dependent on dispatcher messaging types
