+++ dispatcher/messaging/event.go
@@ -1,11 +1,16 @@
 package messaging
 
-// Event the basic event data format
+//KeyValuePair is a key value pair
+type KeyValuePair struct {
+	Key   string      `bson:"key" json:"key"`
+	Value interface{} `bson:"value" json:"value"`
+}
+
+//Event the basic event data format
 type Event struct {
-	ID             string            `json:"id"`
-	Type           string            `json:"type"`
-	PreviousStages []string          `json:"previousStages"`
-	ParentEventID  string            `json:"parentId"`
-	CorrelationID  string            `json:"correlationId"`
-	Data           map[string]string `json:"data"`
+	EventID        string         `json:"eventID"`
+	Type           string         `json:"type"`
+	PreviousStages []string       `json:"previousStages"`
+	CorrelationID  string         `json:"correlationID"`
+	Data           []KeyValuePair `json:"data"`
 }
