+++ sidecar/app/app.go
@@ -13,6 +13,7 @@ import (
 	"github.com/sirupsen/logrus"
 
 	"github.com/gorilla/mux"
+	"github.com/lawrencegripper/ion/dispatcher/messaging"
 	"github.com/lawrencegripper/ion/sidecar/types"
 	log "github.com/sirupsen/logrus"
 )
@@ -305,7 +306,7 @@ func (a *App) commitMeta(metadataPath string) error {
 	if err != nil {
 		return fmt.Errorf("failed to read metadata document '%s' with error '%+v'", metadataPath, err)
 	}
-	var m []types.KeyValuePair
+	var m []messaging.KeyValuePair
 	err = json.Unmarshal(bytes, &m)
 	if err != nil {
 		return fmt.Errorf("failed to unmarshal metadata '%s' with error: '%+v'", metadataPath, err)
@@ -341,7 +342,7 @@ func (a *App) commitEvents(eventsPath string) error {
 			return fmt.Errorf("failed to read file '%s' with error: '%+v'", fileName, err)
 		}
 		// Deserialize event into map
-		var kvps []types.KeyValuePair
+		var kvps []messaging.KeyValuePair
 		decoder := json.NewDecoder(f)
 		err = decoder.Decode(&kvps)
 		if err != nil {
@@ -395,7 +396,7 @@ func (a *App) commitEvents(eventsPath string) error {
 
 		// Create new event
 		eventID := NewGUID()
-		event := types.Event{
+		event := messaging.Event{
 			PreviousStages: []string{},
 			EventID:        eventID,
 			Type:           eventType,
